# Optimized the jeepers out of the scoreboard queries 