--- a/app/models/application_record.rb
+++ b/app/models/application_record.rb
@@ -15,6 +15,12 @@ class ApplicationRecord < ActiveRecord::Base
       end
     end
 
+    # Arel node for a MySQL function that returns 0 if the expression is NULL
+    # Useful for e.g. doing math with a non-existant Quiz Score.
+    def null_as_zero(expression)
+      Arel::Nodes::NamedFunction.new('IFNULL', [cast_to_node(expression), 0])
+    end
+
     # Arel node for a DENSE_RANK window function, can be used in a `SELECT` clause.
     def top?(*orders, top: 20)
       window = Arel::Nodes::Window.new
@@ -22,12 +28,6 @@ class ApplicationRecord < ActiveRecord::Base
       Arel::Nodes::Over.new(Arel::Nodes::NamedFunction.new('DENSE_RANK', []), window).then do |wf|
         wf.lteq(top).as("top_#{top}")
       end
-    end
-
-    # Arel node for a MySQL function that returns 0 if the expression is NULL
-    # Useful for e.g. doing math with a non-existant Quiz Score.
-    def null_as_zero(expression)
-      Arel::Nodes::NamedFunction.new('IFNULL', [cast_to_node(expression), 0])
     end
   end
 end