--- a/app/models/circle_check_score.rb
+++ b/app/models/circle_check_score.rb
@@ -14,11 +14,22 @@ class CircleCheckScore < ApplicationRecord
 
   TOTAL_DEFECTS_DEFAULT = 5
 
+  def self.score_calculation
+    total_defects = arel_table[:total_defects]
+    defects_found = arel_table[:defects_found]
+    scale = Arel::Nodes::Division.new(50, total_defects)
+    null_as_zero Arel::Nodes::Multiplication.new(scale, defects_found)
+  end
+
+  def self.with_scores
+    select score_calculation.as('circle_check_score')
+  end
+
   def as_json(options = {})
     super(options).merge(score: score)
   end
 
   def score
-    (50 / total_defects) * defects_found
+    attributes['circle_check_score'] || (50 / total_defects.to_f) * defects_found
   end
 end