--- a/app/models/circle_check_score.rb
+++ b/app/models/circle_check_score.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 class CircleCheckScore < ApplicationRecord
+  include NormalizedScore
   include ScoreboardPublisher
 
   belongs_to :participant
@@ -14,22 +15,9 @@ class CircleCheckScore < ApplicationRecord
 
   TOTAL_DEFECTS_DEFAULT = 5
 
-  def self.score_calculation
-    total_defects = arel_table[:total_defects]
-    defects_found = arel_table[:defects_found]
-    scale = Arel::Nodes::Division.new(50, total_defects)
-    null_as_zero Arel::Nodes::Multiplication.new(scale, defects_found)
-  end
-
-  def self.with_scores
-    select score_calculation.as('circle_check_score')
-  end
+  def self.score_components = %i[defects_found total_defects]
 
   def as_json(options = {})
     super(options).merge(score: score)
-  end
-
-  def score
-    attributes['circle_check_score'] || ((50 / total_defects.to_f) * defects_found)
   end
 end