--- a/app/models/maneuver_participant.rb
+++ b/app/models/maneuver_participant.rb
@@ -34,11 +34,9 @@ class ManeuverParticipant < ApplicationRecord
   end
 
   def self.scoreboard_grouping
-    grouping = {}
-    ManeuverParticipant.all.group_by(&:participant_id).each_pair do |pid, mps|
-      grouping[pid] = mps.group_by(&:maneuver_id)
+    all.group_by(&:participant_id).transform_values do |mps|
+      mps.group_by(&:maneuver_id).transform_values(&:first)
     end
-    grouping
   end
 
   private