--- a/app/models/null_participant.rb
+++ b/app/models/null_participant.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+class NullParticipant < Participant
+  after_initialize :readonly!
+
+  def number = 0
+  def show_name? = false
+end