--- a/app/models/null_participant.rb
+++ b/app/models/null_participant.rb
@@ -4,5 +4,6 @@ class NullParticipant < Participant
   after_initialize :readonly!
 
   def number = 0
+
   def show_name? = false
 end