--- a/app/models/quiz_score.rb
+++ b/app/models/quiz_score.rb
@@ -16,11 +16,22 @@ class QuizScore < ApplicationRecord
 
   TOTAL_POINTS_DEFAULT = 100
 
+  def self.score_calculation
+    total_points = arel_table[:total_points]
+    points_achieved = arel_table[:points_achieved]
+    scale = Arel::Nodes::Division.new(50, total_points)
+    null_as_zero Arel::Nodes::Multiplication.new(scale, points_achieved)
+  end
+
+  def self.with_scores
+    select score_calculation.as('quiz_score')
+  end
+
   def as_json(options = {})
     super(options).merge(score: score)
   end
 
   def score
-    (50 / total_points * points_achieved).round 1
+    attributes['quiz_score'] || (50 / total_points * points_achieved).round(1)
   end
 end