--- a/app/views/participants/_participant.haml
+++ b/app/views/participants/_participant.haml
@@ -1,4 +1,3 @@
-- scores = @scores[participant.id]
 %tr{data: {participant_id: participant.id}}
   %td.participant{data: {text: participant.number}}
     - if participant.show_name?
@@ -7,7 +6,7 @@
       = participant.display_information(:number)
 
   - @maneuvers.each do |maneuver|
-    - record = scores.try(:[], maneuver.id).try :first
+    - record = participant.maneuver_participants.find_by(maneuver: maneuver)
     %td.maneuver-participant{data: {maneuver_id: maneuver.id}.merge(score_data record)}
       = score_cell record,
         allowed: @can_edit_scores,