--- a/app/views/participants/scoreboard.haml
+++ b/app/views/participants/scoreboard.haml
@@ -1,4 +1,4 @@
-- if Participant.numbered.any?
+- if @participants_exist
   .scoreboard
     %button.btn.btn-primary.fullscreen
       %i.fas.fa-expand