… coverage/.last_run.json
… lib/scoreboard_renderer.rb
… spec/models/application_record_spec.rb
… spec/models/circle_check_score_spec.rb
… spec/models/concerns/normalized_score.rb
… spec/models/maneuver_participant_spec.rb
… spec/models/null_participant_spec.rb
… spec/models/participant_spec.rb
… spec/models/quiz_score_spec.rb