Closes #156

Made the scoreboard ~100x faster

Before:

```
Completed 200 OK in 11389ms (Views: 10950.1ms | ActiveRecord: 171.9ms | Allocations: 7465341)
```

After:

```
Completed 200 OK in 193ms (Views: 123.9ms | ActiveRecord: 4.7ms | Allocations: 199296)
```